New glTF-based model loading
We don't want students to use the MSU-owned hydra/chimera/whatever related formats, because they are undocumented and maintained by god knows who.

Instead, we will use glTF, an open standard with a nice human-readable spec and lots of tools, information and tutorials available online, as well as, well, CC0 scenes. This should make it a lot easier to tinker with scenes, especially for people working on theses in graphics.

This commit adds a new SceneManager, first and foremost. It's a simple class that converts a glTF model to engine-appropriate format and uploads it to the GPU. It uses tinygltf as the parser.

The shadowmap sample was updated to use it. Also a bunch of small fixes that I don't care to put into different commits.

+++ src/render/ImGuiRender.cpp
@@ -82,7 +82,7 @@ void ImGuiRender::InitImGui(vk::Format a_target_format)
       .sType                   = VK_STRUCTURE_TYPE_PIPELINE_RENDERING_CREATE_INFO,
       .pNext                   = nullptr,
       .viewMask                = 0,
-      .colorAttachmentCount    = targetFormats.size(),
+      .colorAttachmentCount    = static_cast<std::uint32_t>(targetFormats.size()),
       .pColorAttachmentFormats = targetFormats.data(),
       .depthAttachmentFormat   = {},
       .stencilAttachmentFormat = {},
